fix(dlm): task timeout was not effective (#2651)
* several bugfix

* fix NPE

+++ server/odc-service/src/main/java/com/oceanbase/odc/service/dlm/model/DataArchiveParameters.java
@@ -64,7 +64,7 @@ public class DataArchiveParameters implements TaskParameters {
 
     private int scanBatchSize;
 
-    private int timeoutMillis;
+    private Long timeoutMillis;
 
     private Set<DBObjectType> syncTableStructure = new HashSet<>();
